refactor: reorganize new files to be less chaotic
also improve keyboard support a bit more using focus/blur events
