+++ app/components/downloads/pack-preview.vue
